Updated Pronoun accesor
this new system is much easier to understand and a far more scaleable piece of code. It is now able to extract the people in groups of nouns so that they may be referenced if they are Unique. Ex "Carlos and Carolina", here "he" refers to Carlos and "she" to Carolina. but in "Carlos and Eric" you don't know to whom "he" refers to.